offscreen-canvas/pixel-manipulation/2d.imageData.object.set.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.imageData.object.set</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>

<h1>2d.imageData.object.set</h1>
<p class="desc">ImageData.data can be modified</p>


<script>
promise_test(function(t) {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

var imgdata = ctx.getImageData(0, 0, 10, 10);
imgdata.data[0] = 100;
_assertSame(imgdata.data[0], 100, "imgdata.data[\""+(0)+"\"]", "100");
imgdata.data[0] = 200;
_assertSame(imgdata.data[0], 200, "imgdata.data[\""+(0)+"\"]", "200");

return Promise.resolve();
}, "ImageData.data can be modified");
</script>
